requirements.txt
pathlib
python-can
canopen
numpy
